+++ src/Nethermind/Nethermind.Blockchain.Test/Visitors/DbBlocksLoaderTests.cs
@@ -34,7 +34,7 @@ public class DbBlocksLoaderTests
             MemDb headersDb = new();
 
             BlockTree testTree = Build.A.BlockTree(genesisBlock).OfChainLength(chainLength).TestObject;
-            for (int i = 0; i < testTree.Head!.Number + 1; i++)
+            for (ulong i = 0ul; i < testTree.Head!.Number + 1ul; i++)
             {
                 Block ithBlock = testTree.FindBlock(i, BlockTreeLookupOptions.None)!;
                 blockStore.Insert(ithBlock);
@@ -80,7 +80,7 @@ public class DbBlocksLoaderTests
             MemDb headersDb = new();
 
             BlockTree testTree = Build.A.BlockTree(genesisBlock).OfChainLength(chainLength).TestObject;
-            for (int i = 0; i < testTree.Head!.Number + 1; i++)
+            for (ulong i = 0ul; i < testTree.Head!.Number + 1ul; i++)
             {
                 Block ithBlock = testTree.FindBlock(i, BlockTreeLookupOptions.None)!;
                 blockStore.Insert(ithBlock);
@@ -170,7 +170,7 @@ public class DbBlocksLoaderTests
         DbBlocksLoader loader = new(tree2, LimboNoErrorLogger.Instance, null, 1);
         await tree2.Accept(loader, tokenSource.Token);
 
-        Assert.That(tree2.BestKnownNumber, Is.EqualTo(3L), "best known");
+        Assert.That(tree2.BestKnownNumber, Is.EqualTo(3ul), "best known");
         BlockTestAssertions.AssertBlockHeaderEquivalent(tree2.Head!.Header, block3B.Header);
         BlockTestAssertions.AssertBlockHeaderEquivalent(tree2.BestSuggestedHeader, block3B.Header);
 
@@ -232,7 +232,7 @@ public class DbBlocksLoaderTests
 
         /* note the block tree historically loads one less block than it could */
 
-        Assert.That(tree2.BestKnownNumber, Is.EqualTo(0L), "best known");
+        Assert.That(tree2.BestKnownNumber, Is.EqualTo(0ul), "best known");
         Assert.That(tree2.Head!.Hash, Is.EqualTo(block0.Hash), "head");
         Assert.That(tree2.BestSuggestedHeader!.Hash, Is.EqualTo(block0.Hash), "suggested");
